Crate fedimint_mint

Re-exports

pub use common::BackupRequest;

pub use common::SignedBackupRequest;

Modules

common

config

db

Macros

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

Structs

BlindNonce

Nonce but blinded by the user key

Endpoint ğŸ”’

Endpoint ğŸ”’

IssuanceData ğŸ”’

Mint

Data structures taking into account different amount tiers Federated mint member mint

MintConsensusItem

A consenus item from one of the federation members contributing partials signatures to blind nonces submitted in it

MintGen

MintGenParams

MintInput

MintOutput

MintOutputBlindSignatures

Result of Federation members confirming MintOutput by contributing partial signatures via MintConsensusItem

MintOutputOutcome

MintOutputSignatureShare

Blind signature share from one Federation peer for a single MintOutput

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

VerifiedNotes

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

Enums

CombineError

MintError

PeerErrorType

__Field ğŸ”’

__Field ğŸ”’

Constants

DEFAULT_MAX_NOTES_PER_DENOMINATION ğŸ”’

By default, the maximum notes per denomination when change-making for users

FIELDS ğŸ”’

FIELDS ğŸ”’

KIND ğŸ”’

Statics

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

CALLSITE ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

META ğŸ”’

Opaque Types

ğŸ”’

ğŸ”’

ğŸ”’