Module fedimint_mint::db

Macros

try 🔒

try 🔒

try 🔒

try 🔒

try 🔒

try 🔒

try 🔒

try 🔒

Structs

DbKeyPrefixIter

An iterator over the variants of Self

ECashUserBackupSnapshot

User’s backup, received at certain time, containing encrypted payload

EcashBackupKey

Key used to store user’s ecash backups

EcashBackupKeyPrefix

MintAuditItemKeyPrefix

NonceKey

NonceKeyPrefix

OutputOutcomeKey

Transaction id and output index identifying an output outcome

OutputOutcomeKeyPrefix

ProposedPartialSignatureKey

ProposedPartialSignaturesKeyPrefix

ReceivedPartialSignatureKey

ReceivedPartialSignatureKeyOutputPrefix

ReceivedPartialSignaturesKeyPrefix

__DeserializeWith 🔒

__DeserializeWith 🔒

__FieldVisitor 🔒

__SerializeWith 🔒

__Visitor 🔒

Enums

DbKeyPrefix

MintAuditItemKey

Represents the amounts of issued (signed) and redeemed (verified) coins for auditing

__Field 🔒

Constants

FIELDS 🔒