Crate fedimint_mint

Modules

config

db 🔒

Structs

BlindNonce

Nonce but blinded by the user key

Mint

Data structures taking into account different amount tiers Federated mint member mint

MintShareErrors

Represents an array of mint indexes that delivered faulty shares

Nonce

Unique ID of a mint note.

Note

An verifiable one time use IOU from the mint.

PartialSigResponse

Blind signature share for a SignRequest

PartiallySignedRequest

SigResponse

Blind signature for a SignRequest

SignRequest

Request to blind sign a certain amount of coins

VerificationCache

Enums

CombineError

MintError

PeerErrorType